+++ CHANGELOG.md
@@ -11,6 +11,13 @@ This change log follows standards by [Keep a CHANGELOG](http://keepachangelog.co
 - **require Node.js v5.x**
 - bump htmlmin version
 - move comment about generator version from Gruntfile to README
+- moved assets including into Gruntfile instead of processing the HTML with `grunt-usemin`, all new files should be now included in the `concat` task according to the comments
+
+### Added
+- asynchronous loading of `main.js`
+
+### Removed
+- `grunt-usemin`
 
 ## [0.9.0] - 2016-03-24
 ### Added